channeld: extract HTLC trim logic into common.
