+++ channeld/test/Makefile
@@ -11,6 +11,7 @@ CHANNELD_TEST_COMMON_OBJS :=			\
 	common/amount.o				\
 	common/daemon_conn.o			\
 	common/htlc_state.o			\
+	common/htlc_trim.o			\
 	common/htlc_tx.o			\
 	common/initial_commit_tx.o		\
 	common/key_derive.o			\
